backend/requirements.txt
tensorflow>=2.13.0
numpy>=1.24.0
Pillow>=10.0.0
scikit-learn>=1.3.0
joblib>=1.3.0
flask>=2.3.0
flask-cors>=4.0.0
werkzeug>=2.3.0
matplotlib>=3.7.0
seaborn>=0.12.0
google-generativeai>=0.3.0
python-dotenv>=1.0.0
requests>=2.31.0
pyTelegramBotAPI>=4.0.0
